Ard py3 (#9)
* Correct import order.

* Update to modern dict api.

* Flip import order.

* Write python 3 version of vnc password get/set funcs.

It's super confusing going between py2/py3 bytes, so I opted to just
separate the funcs out with a python version conditional check. When it
comes time to drop py2, it will be very easy to just snip snip them out.

* Remove outdated service check. Fix agent test.

* Update ard.managed state to work correctly.

- Swap out use of plist.read/write_keys to use prefs.list/set. Writing
to the plist does not affect state immediately like using a CFPrefs
method.
- The state did not report on changes to preferences that were set; only
a subset of changes were listed as changes. This has been corrected.

* Allow ard.managed to use a default `enabled` param of `None`.

This allows admins to not change the activation of the service if they
want to manage other elements (e.g. not allowing VNC).

* Fix dependencies doc.

* Fix bug where missing prefs would cause a lack of action.